+++ tests/Location/LineTest.php
@@ -13,7 +13,7 @@ class LineTest extends TestCase
     public function testCreateLine(): void
     {
         $point1 = new Coordinate(52.5, 13.5);
-        $point2 = new Coordinate(64.1, - 21.9);
+        $point2 = new Coordinate(64.1, -21.9);
 
         $line = new Line($point1, $point2);
 
@@ -24,7 +24,7 @@ class LineTest extends TestCase
     public function testCalculateLength(): void
     {
         $point1 = new Coordinate(52.5, 13.5);
-        $point2 = new Coordinate(64.1, - 21.9);
+        $point2 = new Coordinate(64.1, -21.9);
 
         $line = new Line($point1, $point2);
 
@@ -34,7 +34,7 @@ class LineTest extends TestCase
     public function testGetReverseWorksAsExpected(): void
     {
         $point1 = new Coordinate(52.5, 13.5);
-        $point2 = new Coordinate(64.1, - 21.9);
+        $point2 = new Coordinate(64.1, -21.9);
 
         $line = new Line($point1, $point2);
         $reversedLine = $line->getReverse();
@@ -102,5 +102,51 @@ class LineTest extends TestCase
         $expected = new Bounds(new Coordinate(3.0, 10.0), new Coordinate(-10.0, 20.0));
 
         $this->assertEquals($expected, $line->getBounds());
+    }
+
+    public function testIfGetMidpointWorksAsExpected()
+    {
+        $line = new Line(new Coordinate(0, 0), new Coordinate(10, 20));
+
+        $this->assertEquals(5.07672, $line->getMidpoint()->getLat(), '', 0.0001);
+        $this->assertEquals(9.92267, $line->getMidpoint()->getLng(), '', 0.0001);
+
+        $line = new Line(new Coordinate(1, 1), new Coordinate(-2, -2));
+
+        $this->assertEquals(-0.5, $line->getMidpoint()->getLat(), '', 0.001);
+        $this->assertEquals(-0.5, $line->getMidpoint()->getLng(), '', 0.001);
+
+        $line = new Line(new Coordinate(35, -90), new Coordinate(35.2, -90.4));
+
+        $this->assertEquals(35.1, $line->getMidpoint()->getLat(), '', 0.001);
+        $this->assertEquals(-90.2, $line->getMidpoint()->getLng(), '', 0.001);
+    }
+
+    public function testIfGetMidpointAcrossLongitudeBorderWorksAsExpected()
+    {
+        $line = new Line(new Coordinate(0.0, -179.0), new Coordinate(0.0, 179.0));
+        $this->assertEquals(new Coordinate(0.0, -180.0), $line->getMidpoint());
+
+        $line = new Line(new Coordinate(0.0, -178.0), new Coordinate(0.0, 179.0));
+        $this->assertEquals(new Coordinate(0.0, -179.5), $line->getMidpoint());
+    }
+
+    public function testIfGetIntermediatePointWorksAsExpected()
+    {
+        $line = new Line(new Coordinate(0, 0), new Coordinate(0, 1));
+        $this->assertEquals(new Coordinate(0, 0), $line->getIntermediatePoint(0.));
+        $line = new Line(new Coordinate(0, 0), new Coordinate(0, 1));
+        $this->assertEquals(new Coordinate(0, .5), $line->getIntermediatePoint(.5));
+        $line = new Line(new Coordinate(0, 0), new Coordinate(0, 1));
+        $this->assertEquals(new Coordinate(0, 1), $line->getIntermediatePoint(1.));
+    }
+
+    public function testIfGetIntermediatePointThrowsExceptionForAntipodes()
+    {
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionCode(5382449689);
+
+        $line = new Line(new Coordinate(45, -45), new Coordinate(-45, 135));
+        $line->getIntermediatePoint(.5);
     }
 }